+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+### Changed
+
+### Removed
+
+### Fixed
+- Picker's Marquee will now animate when it's parent is focused. 
+
+
 ## [1.0.0-alpha.4] - 2016-12-2
 
 ### Added